+++ src/index.tsx
@@ -11,3 +11,4 @@ export * from './modules/tokens';
 export * from './modules/checkout';
 
 export * from './modules/dashboard';
+export * from './modules/addFunds';