+++ src/modules/shared/hooks/useProfile/useProfile.ts
@@ -1,4 +1,9 @@
+import { useMutation } from 'react-query';
+
 import { PixwayAPIRoutes } from '../../enums/PixwayAPIRoutes';
+import { W3blockAPI } from '../../enums/W3blockAPI';
+import { useAxios } from '../useAxios';
+import { useCompanyConfig } from '../useCompanyConfig';
 import { useGetW3blockIdSDK } from '../useGetW3blockIdSDK';
 import { usePrivateQuery } from '../usePrivateQuery';
 
@@ -18,4 +23,20 @@ export const useProfile = () => {
   return usePrivateQuery([PixwayAPIRoutes.GET_PROFILE], async () => {
     return (await getW3blockIdSDK()).api.users.getProfileByUserLogged();
   });
+};
+
+export const usePatchProfile = () => {
+  const axios = useAxios(W3blockAPI.ID);
+  const { companyId } = useCompanyConfig();
+
+  const _patchProfile = (name: string) => {
+    return axios.patch(
+      PixwayAPIRoutes.PATCH_PROFILE.replace('{companyId}', companyId),
+      { name }
+    );
+  };
+
+  const patchProfile = useMutation(_patchProfile);
+
+  return patchProfile;
 };